Add body schema validation
