+++ lib/middleware.js
@@ -1,5 +1,12 @@
 import Err from './error.js';
 import Param from './param.js';
+import Ajv from 'ajv';
+import addFormats from 'ajv-formats';
+
+const ajv = addFormats(new Ajv({
+    allErrors: true,
+    removeAdditional: true
+}));
 
 /**
  * @class
@@ -16,7 +23,6 @@ export default class Middleware {
      */
     static param(match, type) {
         return function (req, res, next) {
-
             try {
                 Param[type](req, match.replace(':', ''));
             } catch (err) {
@@ -88,5 +94,52 @@ export default class Middleware {
 
             return next();
         };
+    }
+
+    /**
+     * Express middleware to validate the response body
+     *
+     * @param {Object} schema JSON Schema
+     *
+     * @returns {Function}
+     */
+    static res(schema) {
+        const validate = ajv.compile(schema)
+
+        return function (req, res, next) {
+            var json = res.json;
+            res.json = function(obj) {
+                const jsonres = validate(obj);
+
+                json.call(this, obj);
+            };
+            next();
+        }
+    }
+
+    /**
+     * Convert validation errors into standardized JSON Error Messages
+     */
+    static error() {
+        return function(err, req, res, next) {
+            if (err instanceof ValidationError) {
+                let errs = [];
+                if (err.validationErrors.body) {
+                    errs = errs.concat(err.validationErrors.body);
+                }
+
+                if (err.validationErrors.query) {
+                    errs = errs.concat(err.validationErrors.query);
+                }
+
+                return Err.respond(
+                    new Err(400, null, 'validation error'),
+                    res,
+                    errs
+                );
+            } else {
+                next(err);
+            }
+        }
     }
 }
