+++ lib/schema.js
@@ -150,7 +150,17 @@ export default class Schemas {
 
         flow[1].push(this.validate(opts));
 
+        console.error(schemas);
+        if (schemas.res) flow[1].push(Middleware.res(schemas.res));
+
         return flow;
+    }
+
+    /**
+     * Convert validation errors into standardized JSON Error Messages
+     */
+    error() {
+        this.router.use(Middleware.error());
     }
 
     /**
@@ -184,32 +194,6 @@ export default class Schemas {
                 message: 'API endpoint does not exist!',
                 messages: []
             });
-        });
-    }
-
-    /**
-     * Convert validation errors into standardized JSON Error Messages
-     */
-    error() {
-        this.router.use((err, req, res, next) => {
-            if (err instanceof ValidationError) {
-                let errs = [];
-                if (err.validationErrors.body) {
-                    errs = errs.concat(err.validationErrors.body);
-                }
-
-                if (err.validationErrors.query) {
-                    errs = errs.concat(err.validationErrors.query);
-                }
-
-                return Err.respond(
-                    new Err(400, null, 'validation error'),
-                    res,
-                    errs
-                );
-            } else {
-                next(err);
-            }
         });
     }
