+++ package.json
@@ -19,6 +19,7 @@
     },
     "dependencies": {
         "ajv": "^8.6.3",
+        "ajv-formats": "^2.1.1",
         "body-parser": "^1.20.0",
         "express": "^4.17.1",
         "express-json-validator-middleware": "^3.0.0",
